Se agrega el servicio de vehiculo, funciona todo menos prueba de integracion de craecion de registro

+++ src/main/java/co/com/ceiba/dominio/puerto/repositorio/RepositorioRegistroVehiculo.java
@@ -35,6 +35,6 @@ public interface RepositorioRegistroVehiculo {
 	 * @param vehiculo
 	 * @return si hay cupo o no
 	 */
-	long validarCuposPorTipoVehiculo(String tipo);
+	int validarCuposPorTipoVehiculo(String tipo);
 	
 }
